Solving property deletion and right option repetitions bugs
